+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 3.10.0 - 2018-07-26
 - Change return type of `getIterator` method to generic `iterable`
 - Add `IEnumerable` interface to implement `IteratorAggregate` and `Countable`
 - Add `PrioritizedCollection` to store values by priority